add confirmation prompt for account deactivation

+++ src/public/js/profile.js
@@ -48,22 +48,26 @@ $( "#deleteBtn" ).click( function(e) {
    var url = 'http://localhost:3000/profile/' + id;
    console.log("clicked to delete " + id);
    console.log(url);
-
-   $.ajax({
-    url: url,
-   	type: 'DELETE',
-    success: function() {
-      console.log('success');
-    },
-    error: function (e) {
+  if (confirm("Are you sure you want to deactivate your account?")) {
+    $.ajax({
+      url: url,
+      type: 'DELETE',
+      success: function() {
+        console.log('success');
+      },
+      error: function (e) {
         console.log('error')
         console.log(e);
-    },
-    complete: function() {
-      prompt("Are you sure you want to deactivate your account?")
-    }
-   }).then(function(){
-     console.log('and then');
-     window.location.replace('/signout');
-   });
+      },
+      complete: function() {
+        alert("Account Deactivated!");
+      }
+    }).then(function(){
+      console.log('and then');
+      window.location.replace('/signout');
+    });
+  } else {
+    window.location.reload();
+  }
+
 });
